+++ RandomBalancer/ILoadBalancer.cs
@@ -1,0 +1,7 @@
+namespace RandomBalancer
+{
+    public interface ILoadBalancer<T>
+    {
+        T GetNext();
+    }
+}