+++ RandomBalancer/Server.cs
@@ -1,0 +1,14 @@
+namespace RandomBalancer
+{
+    public class Server
+    {
+        public string Name { get; }
+
+        public Server(string name)
+        {
+            Name = name;
+        }
+
+        public override string ToString() => Name;
+    }
+}